text: Change text measures logic
Before this text was rendering form the bottom line.
However it was broken and it was really hard to figure out where the text
will be actually rendered.
This is the attempt to fix it.

Now text Y will be the base line of the text and measures are calculated from the base line as well.