camera: Remove projection from batcher state
In the first quad-gl implementation camera projection was just one of the
parameters in the quad-gl state. When camera changed - quad-gl checked the new batcher
param and started a new drawcall.

But now when camera change we always flush all the drawcalls to the screen - to make sure that
all rendertargets are OK.

So keeping projection as a part of the batcher does not really make sense anymore.

And on top of this static batcher is going to be refactored and it should be possible to
batch geometry once and than render it into multiple cameras. It is just a plan now, but
those plans are part of the motivation behind this change.