Allow users to use arbitrary futures (#288)
exec: Allow users to use arbitrary futures

This change removes old and very unsafe future's context transmute and use Wakers instead. 
Now it is possible to `.await` third-party, not macroquad futures from macroquad's main loop or coroutines.
It is still impossible to run macroquad's main loop from third-party executor like tokio. But this is intended and there going to be a nice and descriptive panic instead of weird segfault we had before.