Explore fixing window clipping on high_dpi
In #390 I demonstrated that on a high DPI screen the label for a window
gets cut off very abruptly. I realised that it was being cut off at
roughly half the window size, which made me suspect that a lack of
consideration for the DPI might be to blame.

I went diving into the code and came across the code that clips. As an
experiment I tried multipling the values of the clip rect by the DPI
value, and it worked!

Now, I **do not** suspect this is the right solution; I think the
solution might be down in miniquad, but to be honest I've reached the
limits of my very much beginner Rust knowledge. I'm mostly sharing this
PR for info and in case it helps figure out the real source of the
issue.

If I have some time I'll try to dive into miniquad - any pointers on
debugging this welcome!