Fix: incorrect text dimensions from measure_text on custom dpi
Why:
The dpi scaling is not applied in the measure_text function to resulting text dimensions, but in the draw_text_ex it does.

How:
Just apply a dpi scaling in the measure_text function by analogy with the draw_text_ex function.

Issue: https://github.com/not-fl3/macroquad/issues/373