round get_fps()
Before the decimal was just being dropped, causing the values of get_fps() to always oscillate between 59 and 60, for example. This isn't very helpful when debugging, and it's due to the decimal being cut off even if it's, let's say, 59.9 due to floating point arithmetic.

Now the value is rounded, which is more helpful when monitoring for massive frame drops.